backend/rest-api/login-user/services/auth_service.py
import bcrypt
import jwt
import datetime
import os
from dotenv import load_dotenv
from config.database import get_db_connection

load_dotenv()
SECRET_KEY = os.getenv("JWT_SECRET_KEY")

def hash_password(password):
    return bcrypt.hashpw(password.encode('utf-8'), bcrypt.gensalt()).decode('utf-8')

def verify_password(password, hashed_password):
    return bcrypt.checkpw(password.encode('utf-8'), hashed_password.encode('utf-8'))

def generate_token(user):
    payload = {
        "id": user["id"],
        "email": user["email"],
        "role": user["role"],
        "exp": datetime.datetime.utcnow() + datetime.timedelta(hours=1)
    }
    return jwt.encode(payload, SECRET_KEY, algorithm="HS256")
